rename: MVC 방식의 폴더 구조화
