+++ src/main/java/lotto/domain/Lotto.java
@@ -1,6 +1,7 @@
-package lotto;
+package lotto.domain;
 
 import java.util.List;
+import lotto.validator.Validator;
 
 public class Lotto {
     private final List<Integer> numbers;
@@ -11,10 +12,6 @@ public class Lotto {
     }
 
     private void validate(List<Integer> numbers) {
-        if (numbers.size() != 6) {
-            throw new IllegalArgumentException("[ERROR] 로또 번호는 6개여야 합니다.");
-        }
+        Validator.validateWinningNumber(numbers);
     }
-
-    // TODO: 추가 기능 구현
 }
